refactor(scripts): extract shared segmentation sweep runner
run_segmentation_representative_sweep.py and run_segmentation_protocol_study.py
shared ~145 identical lines: near-identical runner classes plus duplicate
Model/Job dataclasses, GPU parsing, and path resolution. Both now build on
scripts/_seg_sweep_common.py and keep only their job construction and CLI.
Both --help outputs are unchanged.
